requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
pyyaml
openenv-core
jinja2
requests
httpx
huggingface-hub
